+++ arduino_due/ex_tft_ssd1963/ex_tft_ssd1963.cpp
@@ -178,7 +178,7 @@ static BOOL actPen()
 	return !digitalRead(PIN_TS_PEN);
 }
 
-static void actTransfer(PifSpiDevice *p_owner, uint8_t* p_write, uint8_t* p_read, uint16_t size)
+static void actTransfer(PifSpiDevice *p_owner, uint8_t* p_write, uint8_t* p_read, size_t size)
 {
 	uint16_t i;
