+++ arduino_mega_2560/ex_storage_var_i2c/ex_storage_var_i2c.cpp
@@ -73,7 +73,7 @@ static void evtLedToggle(void* p_issuer)
 	sw ^= 1;
 }
 
-static PifI2cReturn actI2cRead(PifI2cDevice *p_owner, uint32_t iaddr, uint8_t isize, uint8_t* p_data, uint16_t size)
+static PifI2cReturn actI2cRead(PifI2cDevice *p_owner, uint32_t iaddr, uint8_t isize, uint8_t* p_data, size_t size)
 {
 #ifdef USE_I2C_WIRE
 	int i;
@@ -106,7 +106,7 @@ static PifI2cReturn actI2cRead(PifI2cDevice *p_owner, uint32_t iaddr, uint8_t is
     return IR_COMPLETE;
 }
 
-static PifI2cReturn actI2cWrite(PifI2cDevice *p_owner, uint32_t iaddr, uint8_t isize, uint8_t* p_data, uint16_t size)
+static PifI2cReturn actI2cWrite(PifI2cDevice *p_owner, uint32_t iaddr, uint8_t isize, uint8_t* p_data, size_t size)
 {
 #ifdef USE_I2C_WIRE
 	int i;
