+++ nucleo_f767zi/exPmlcdI2c/Core/Src/main.c
@@ -114,7 +114,7 @@ static void actLedLState(PifId usPifId, uint32_t unState)
   HAL_GPIO_WritePin(GPIOB, LD2_Pin, unState & 1);
 }
 
-static PifI2cReturn actI2cWrite(PifI2cDevice *p_owner, uint32_t iaddr, uint8_t isize, uint8_t* p_data, uint16_t size)
+static PifI2cReturn actI2cWrite(PifI2cDevice *p_owner, uint32_t iaddr, uint8_t isize, uint8_t* p_data, size_t size)
 {
 	if (isize) {
 		return (HAL_I2C_Mem_Write_IT(&hi2c1, p_owner->addr << 1, iaddr, isize, p_data, size) == HAL_OK) ? IR_WAIT : IR_ERROR;
